split a file
