+++ enclone_main/src/lib.rs
@@ -1,4 +1,5 @@
 // Copyright (c) 2021 10x Genomics, Inc. All rights reserved.
 
 pub mod main_enclone;
+pub mod setup;
 pub mod split_orbits;
